+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-03T05:48:15.900Z",
+    "2026-07-03T08:32:23.196Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {